+++ CHANGELOG.md
@@ -7,12 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
 -   Test webpack@next
 -   Migrate to Typescript
 -   Update packages
 -   Use [temp-sandbox](https://github.com/chrisblossom/temp-sandbox) for tests
 -   Use `mode: 'development'` for testing webpack >= 4
 -   verbose mode log filename relative to webpack's output path
+
+### Fixed
+
+-   Include dot files in pattern matching
 
 ## [2.0.3] - 2018-07-25